elfutils: remove i386_dis.h/x86_64_dis.h compilation targets
Since we provide those files manually (i386_gendis, which is needed for
generating those files, has to be run on host and would fail when
compiling for other architectures), the mentioned compilation targets
in libcpu/ are not needed anymore.

This change will avoid a nasty race condition when running "make -jX
install" resulting in a zero size libebl_i386.so file. The issue happens
because, at "make install" time, the *_dis.h prerequisites will be newer
than the target itself, triggering a chain of recompilations while, in
the same time, the binary files are copied to the destination directory.
Hence, the zero sized file...

[YOCTO #4131]

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>