libdrm: use RRECOMMENDS for libdrm-drivers
As libdrm enables modules automatically, based on what it needs / can build,
it's unlikely we'll *ever* get a situation where all the modules are emitted,
so libdrm-drivers, which depends on all of them, ends up being useless and
non-installable. Make it use rrecommends, so it only pulls in what we have.

Signed-off-by: Christopher Larson <kergoth@gmail.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>