multilib_header: Do not install mutlilib headers for musl
musl is not multilib and this creates trouble. eg. when
util-linux probes for ncurses it does not find it because
ncurses has installed the multilibbed header and this
header includes bits/wordsize.h and this header does not
exist on musl systems. If and when musl adds multilib
support we will revisit it.

Signed-off-by: Khem Raj <raj.khem@gmail.com>